+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - **github** support multiple environments (production, preview etc)
 - **github** remote interception of detected changes (schema check) via http
 - **github** managing unrelated PRs separately
+- **action** use original SHA, not SHA from `pull_request` event [PR #1440](https://github.com/kamilkisiela/graphql-inspector/pull/1440)
+- **github** **action** assume valid schema to avoid missing directive definitions [PR #1440](https://github.com/kamilkisiela/graphql-inspector/pull/1440)
 
 ### v1.30.4